Kill configuration sets concept
Configuration Sets (namespacing) were introduced to
allow storage of multiple configurations, and specify
a single active configuration at any one time.

The vision was that it would allow you to easily switch
between different configurations like "weekday" vs "weekend".

However, it adds a lot of complexity to the codebase,
(as you can see in the size of this diff)
and overhead in a number of extra redis calls
(from 34 to 12 to get current priorities). There
is no clear advantage of managing configuration sets
within qtrix itself. Whatever process would be activating
new configuration sets could just as easily specify
new queue weights and overrides. How the configurations
are stored is irrelevant to qtrix.