Fix pruning under ruby 2.x
Pruning currently relies on deleting from redis using
an exact string match of the marshal'ed representation
of a matrix Row.

Unfortunately, the marhsal'ed representation of a
BigDecimal can vary on ruby 2.x. It is not clear if this
is a ruby bug, but it definitely gets in our way.
https://gist.github.com/joshuaflanagan/44a8c4f3d8cf53b24e60

For now, we'll store the `resource_percentage` as a
float instead of a BigDecimal. In the future, we may
stop persisting the percentage (since it should be calculable
from the other values), and/or improve matrix storage so
that it is not dependent on exact match of marshal'ed strings.