Make Qtrix::Queue immutable
Tell the Queue its relative weight on creation, instead
of having it calculate it itself (which requires knowledge
of all the other Queues).

This makes it a little easier to reason about, but also
eliminates an n+1 redis call problem. Each queue used to
fetch the entire list from redis when it needed to calculate
its own relative weight.